feat: adiciona centralizador de erros, validações no service e testes para os services existentes no backend

+++ backend/src/main/java/io/turismo/backend/service/SpotManagerService.java
@@ -4,10 +4,15 @@ import io.turismo.backend.dto.spot_manager.SpotManagerCreateDTO;
 import io.turismo.backend.dto.spot_manager.SpotManagerDTO;
 import io.turismo.backend.dto.spot_manager.SpotManagerSimpleDTO;
 import io.turismo.backend.dto.spot_manager.SpotManagerUpdateDTO;
+import io.turismo.backend.exception.ObjectNotFoundException;
+import io.turismo.backend.exception.UserIsNotAdminOrOwnerException;
+import io.turismo.backend.exception.UserIsNotOwnerException;
 import io.turismo.backend.mapper.SpotManagerMapper;
 import io.turismo.backend.model.SpotManager;
+import io.turismo.backend.model.User;
 import io.turismo.backend.model.enums.UserRole;
 import io.turismo.backend.repository.SpotManagerRepository;
+import io.turismo.backend.repository.UserRepository;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import java.util.UUID;
@@ -17,12 +22,14 @@ public class SpotManagerService{
     private final SpotManagerRepository spotManagerRepository;
     private final SpotManagerMapper spotManagerMapper;
     private final UserService userService;
+    private final UserRepository userRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
-    public SpotManagerService(SpotManagerRepository spotManagerRepository, SpotManagerMapper spotManagerMapper, UserService userService, BCryptPasswordEncoder bCryptPasswordEncoder) {
+    public SpotManagerService(SpotManagerRepository spotManagerRepository, SpotManagerMapper spotManagerMapper, UserService userService, UserRepository userRepository, BCryptPasswordEncoder bCryptPasswordEncoder) {
         this.spotManagerRepository = spotManagerRepository;
         this.spotManagerMapper = spotManagerMapper;
         this.userService = userService;
+        this.userRepository = userRepository;
         this.bCryptPasswordEncoder = bCryptPasswordEncoder;
     }
 
@@ -42,20 +49,24 @@ public class SpotManagerService{
     public SpotManagerSimpleDTO getSpotManager(UUID spotManagerId) {
         return spotManagerMapper.toSimpleDTO(
                 spotManagerRepository.findById(spotManagerId)
-                        .orElseThrow(() -> new RuntimeException("Gerente não encontrado"))
+                        .orElseThrow(() -> new ObjectNotFoundException("Gerente não encontrado"))
         );
     }
 
     public SpotManagerDTO currentSpotManager(UUID spotManagerId) {
         SpotManager spotManager = spotManagerRepository.findById(spotManagerId)
-                .orElseThrow(() -> new RuntimeException("Gerente não encontrado"));
+                .orElseThrow(() -> new ObjectNotFoundException("Gerente não encontrado"));
 
         return spotManagerMapper.toDTO(spotManager);
     }
 
-    public SpotManagerSimpleDTO updateSpotManager(SpotManagerUpdateDTO spotManagerUpdateDTO, UUID spotManagerId){
+    public SpotManagerSimpleDTO updateSpotManager(SpotManagerUpdateDTO spotManagerUpdateDTO, UUID spotManagerId, UUID userId){
         SpotManager spotManager = spotManagerRepository.findById(spotManagerId)
-                .orElseThrow(() -> new RuntimeException("Gerente não encontrado"));
+                .orElseThrow(() -> new ObjectNotFoundException("Gerente não encontrado"));
+
+        if(!spotManager.getUser().getId().equals(userId)) {
+            throw new UserIsNotOwnerException("Você não tem autorização para isso");
+        }
 
         spotManagerMapper.updateEntityFromDto(spotManagerUpdateDTO, spotManager);
 
@@ -64,9 +75,15 @@ public class SpotManagerService{
         return spotManagerMapper.toSimpleDTO(spotManagerUpdated);
     }
 
-    public void deleteSpotManager(UUID spotManagerId){
+    public void deleteSpotManager(UUID spotManagerId, UUID userId){
         SpotManager spotManager = spotManagerRepository.findById(spotManagerId)
-                .orElseThrow(() -> new RuntimeException("Gerente não encontrado"));
+                .orElseThrow(() -> new ObjectNotFoundException("Gerente não encontrado"));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new ObjectNotFoundException("Usuário não encontrado"));
+
+        if(!spotManager.getUser().getId().equals(userId) && !user.getRole().equals(UserRole.ADMIN)) {
+            throw new UserIsNotAdminOrOwnerException("Você não tem autorização para isso");
+        }
 
         spotManagerRepository.delete(spotManager);
     }
