+++ backend/src/main/java/io/turismo/backend/service/TourGuideService.java
@@ -3,10 +3,16 @@ package io.turismo.backend.service;
 import io.turismo.backend.dto.tour_guide.TourGuideCreateDTO;
 import io.turismo.backend.dto.tour_guide.TourGuideDTO;
 import io.turismo.backend.dto.tour_guide.TourGuideUpdateDTO;
+import io.turismo.backend.exception.ObjectAlreadyExistsException;
+import io.turismo.backend.exception.ObjectNotFoundException;
+import io.turismo.backend.exception.UserIsNotAdminOrOwnerException;
+import io.turismo.backend.exception.UserIsNotOwnerException;
 import io.turismo.backend.mapper.TourGuideMapper;
 import io.turismo.backend.model.TourGuide;
+import io.turismo.backend.model.User;
 import io.turismo.backend.model.enums.UserRole;
 import io.turismo.backend.repository.TourGuideRepository;
+import io.turismo.backend.repository.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -18,12 +24,14 @@ public class TourGuideService {
     private final TourGuideRepository tourGuideRepository;
     private final TourGuideMapper tourGuideMapper;
     private final UserService userService;
+    private final UserRepository userRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
-    public TourGuideService(TourGuideRepository tourGuideRepository, TourGuideMapper tourGuideMapper, UserService userService, BCryptPasswordEncoder bCryptPasswordEncoder) {
+    public TourGuideService(TourGuideRepository tourGuideRepository, TourGuideMapper tourGuideMapper, UserService userService, UserRepository userRepository, BCryptPasswordEncoder bCryptPasswordEncoder) {
         this.tourGuideRepository = tourGuideRepository;
         this.tourGuideMapper = tourGuideMapper;
         this.userService = userService;
+        this.userRepository = userRepository;
         this.bCryptPasswordEncoder = bCryptPasswordEncoder;
     }
 
@@ -32,7 +40,7 @@ public class TourGuideService {
         boolean cadasturExists = tourGuideRepository.existsByCadastur(dto.cadastur());
 
         if(cadasturExists) {
-            throw new RuntimeException("Cadastur já cadastrado");
+            throw new ObjectAlreadyExistsException("Cadastur já cadastrado");
         }
 
         TourGuide newTourGuide = tourGuideMapper.toEntity(dto);
@@ -46,31 +54,47 @@ public class TourGuideService {
 
     public TourGuideDTO getTourGuide(UUID tourGuideId) {
         TourGuide tourGuide = tourGuideRepository.findById(tourGuideId)
-                .orElseThrow(() -> new RuntimeException("Guia de Turismo não encontrado"));
+                .orElseThrow(() -> new ObjectNotFoundException("Guia de Turismo não encontrado"));
 
         return tourGuideMapper.toDTO(tourGuide);
-    }
-
-    public TourGuideDTO updateTourGuide(TourGuideUpdateDTO tourGuideUpdate, UUID tourGuideId) {
-        TourGuide tourGuide = tourGuideRepository.findById(tourGuideId)
-                .orElseThrow(() -> new RuntimeException("Guia de Turismo não encontrado"));
-
-        tourGuideMapper.updateEntityFromDto(tourGuideUpdate, tourGuide);
-
-        TourGuide tourGuideUpdated = tourGuideRepository.save(tourGuide);
-
-        return tourGuideMapper.toDTO(tourGuideUpdated);
-    }
-
-    public void deleteTourGuide(UUID tourGuideId) {
-        TourGuide tourGuide = tourGuideRepository.findById(tourGuideId)
-                .orElseThrow(() -> new RuntimeException("Guia de Turismo não encontrado"));
-
-        tourGuideRepository.delete(tourGuide);
     }
 
     public Page<TourGuideDTO> getTourGuidesByTouristSpot(UUID touristSpotId, Pageable pageable) {
         return tourGuideRepository.findAllByTouristSpots_TouristSpotId(touristSpotId, pageable)
                 .map(tourGuideMapper::toDTO);
+    }
+
+    public TourGuideDTO updateTourGuide(TourGuideUpdateDTO tourGuideUpdate, UUID tourGuideId, UUID userId) {
+        TourGuide tourGuide = tourGuideRepository.findById(tourGuideId)
+                .orElseThrow(() -> new ObjectNotFoundException("Guia de Turismo não encontrado"));
+
+        if(!tourGuide.getUser().getId().equals(userId)) {
+            throw new UserIsNotOwnerException("Você não tem autorização para isso");
+        }
+
+        if (tourGuideUpdate.cadastur() != null && !tourGuideUpdate.cadastur().equals(tourGuide.getCadastur())) {
+            boolean cadasturExists = tourGuideRepository.existsByCadastur(tourGuideUpdate.cadastur());
+            if(cadasturExists) {
+                throw new ObjectAlreadyExistsException("Cadastur já cadastrado");
+            }
+        }
+
+        tourGuideMapper.updateEntityFromDto(tourGuideUpdate, tourGuide);
+        TourGuide tourGuideUpdated = tourGuideRepository.save(tourGuide);
+
+        return tourGuideMapper.toDTO(tourGuideUpdated);
+    }
+
+    public void deleteTourGuide(UUID tourGuideId, UUID userId) {
+        TourGuide tourGuide = tourGuideRepository.findById(tourGuideId)
+                .orElseThrow(() -> new ObjectNotFoundException("Guia de Turismo não encontrado"));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new ObjectNotFoundException("Usuário não encontrado"));
+
+        if(!tourGuide.getUser().getId().equals(userId) && !user.getRole().equals(UserRole.ADMIN)) {
+            throw new UserIsNotAdminOrOwnerException("Você não tem autorização para isso");
+        }
+
+        tourGuideRepository.delete(tourGuide);
     }
 }
