+++ backend/src/main/java/io/turismo/backend/service/TouristSpotService.java
@@ -3,13 +3,20 @@ package io.turismo.backend.service;
 import io.turismo.backend.dto.geojson.GeoFeatureCollectionDTO;
 import io.turismo.backend.dto.geojson.GeoFeatureDTO;
 import io.turismo.backend.dto.tourist_spot.*;
+import io.turismo.backend.exception.UserIsNotAdminOrOwnerException;
+import io.turismo.backend.exception.UserIsNotOwnerException;
+import io.turismo.backend.exception.ObjectNotFoundException;
+import io.turismo.backend.exception.InvalidDateException;
 import io.turismo.backend.mapper.TouristSpotMapper;
 import io.turismo.backend.model.City;
 import io.turismo.backend.model.SpotManager;
 import io.turismo.backend.model.TouristSpot;
+import io.turismo.backend.model.User;
+import io.turismo.backend.model.enums.UserRole;
 import io.turismo.backend.repository.CityRepository;
 import io.turismo.backend.repository.SpotManagerRepository;
 import io.turismo.backend.repository.TouristSpotRepository;
+import io.turismo.backend.repository.UserRepository;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.geom.Point;
@@ -17,6 +24,7 @@ import org.locationtech.jts.geom.PrecisionModel;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import java.time.LocalTime;
 import java.util.List;
 import java.util.UUID;
 
@@ -27,22 +35,29 @@ public class TouristSpotService{
     private final SpotManagerRepository spotManagerRepository;
     private final CityRepository cityRepository;
     private final TagService tagService;
+    private final UserRepository userRepository;
 
-    public TouristSpotService(TouristSpotRepository touristSpotRepository, TouristSpotMapper touristSpotMapper, SpotManagerRepository spotManagerRepository, CityRepository cityRepository, TagService tagService) {
+    public TouristSpotService(TouristSpotRepository touristSpotRepository, TouristSpotMapper touristSpotMapper, SpotManagerRepository spotManagerRepository, CityRepository cityRepository, TagService tagService, UserRepository userRepository) {
         this.touristSpotRepository = touristSpotRepository;
         this.touristSpotMapper = touristSpotMapper;
         this.spotManagerRepository = spotManagerRepository;
         this.cityRepository = cityRepository;
         this.tagService = tagService;
+        this.userRepository = userRepository;
     }
 
-    public TouristSpotDTO createTouristSpot(TouristSpotCreateDTO dto, UUID spotManagerId){
-        SpotManager spotManager = spotManagerRepository.findById(spotManagerId)
-                .orElseThrow(() -> new RuntimeException("Gerente não encontrado"));
+    public TouristSpotDTO createTouristSpot(TouristSpotCreateDTO dto, UUID userId){
+        SpotManager spotManager = spotManagerRepository.findByUser_Id(userId)
+                .orElseThrow(() -> new ObjectNotFoundException("Gerente não encontrado"));
         City city = cityRepository.findById(dto.cityId())
-                .orElseThrow(() -> new RuntimeException("Cidade não encontrada"));
+                .orElseThrow(() -> new ObjectNotFoundException("Cidade não encontrada"));
 
         TouristSpot touristSpot = touristSpotMapper.toEntity(dto);
+        
+        if (!timeIsValid(touristSpot.getOpensAt(), touristSpot.getClosesAt())) {
+            throw new InvalidDateException("Horário de fechamento não pode ser antes do horário de abertura");
+        }
+
         touristSpot.setTags(tagService.convertNamesToTags(dto.tags()));
         touristSpot.setCity(city);
         touristSpot.setSpotManager(spotManager);
@@ -58,7 +73,7 @@ public class TouristSpotService{
         return touristSpotMapper.toDTO(
                 touristSpotRepository
                         .findById(touristSpotId)
-                        .orElseThrow(() -> new RuntimeException("Ponto turístico não encontrado"))
+                        .orElseThrow(() -> new ObjectNotFoundException("Ponto turístico não encontrado"))
         );
     }
 
@@ -69,8 +84,7 @@ public class TouristSpotService{
                     var properties = new TouristSpotToMapDTO(spot.getTouristSpotId(), spot.getName());
 
                     return new GeoFeatureDTO<>(properties, spot.getLocation());
-                })
-                .toList();
+                }).toList();
 
         return new GeoFeatureCollectionDTO<>(features);
     }
@@ -105,18 +119,27 @@ public class TouristSpotService{
                 );
     }
 
-    public void updateTouristSpot(UUID touristSpotId, TouristSpotUpdateDTO touristSpotUpdate){
+    public void updateTouristSpot(UUID touristSpotId, TouristSpotUpdateDTO touristSpotUpdate, UUID userId){
         TouristSpot touristSpot = touristSpotRepository.findById(touristSpotId)
-                .orElseThrow(() -> new RuntimeException("Ponto turístico não encontrado"));
+                .orElseThrow(() -> new ObjectNotFoundException("Ponto turístico não encontrado"));
+
+        if (!touristSpot.getSpotManager().getUser().getId().equals(userId)) {
+            throw new UserIsNotOwnerException("Você não tem autorização para isso");
+        }
 
         if(touristSpotUpdate.cityId() != null){
             City city = cityRepository.findById(touristSpotUpdate.cityId())
-                    .orElseThrow(() -> new RuntimeException("Cidade não encontrada"));
+                    .orElseThrow(() -> new ObjectNotFoundException("Cidade não encontrada"));
 
             touristSpot.setCity(city);
         }
 
         touristSpotMapper.updateEntityFromDTO(touristSpotUpdate, touristSpot);
+        
+        if (!timeIsValid(touristSpot.getOpensAt(), touristSpot.getClosesAt())) {
+            throw new InvalidDateException("Horário de fechamento não pode ser antes do horário de abertura");
+        }
+
         touristSpot.setTags(tagService.convertNamesToTags(touristSpotUpdate.tags()));
 
         if (touristSpot.getSocialsMedia() != null) {
@@ -126,10 +149,24 @@ public class TouristSpotService{
         touristSpotRepository.save(touristSpot);
     }
 
-    public void deleteTouristSpot(UUID touristSpotId){
+    public void deleteTouristSpot(UUID touristSpotId, UUID userId){
         TouristSpot touristSpot = touristSpotRepository.findById(touristSpotId)
-                .orElseThrow(() -> new RuntimeException("Ponto turístico não encontrado"));
+                .orElseThrow(() -> new ObjectNotFoundException("Ponto turístico não encontrado"));
+        User user = userRepository.findById(userId)
+                        .orElseThrow(() -> new ObjectNotFoundException("Usuário não existe"));
+        SpotManager spotManager = touristSpot.getSpotManager();
+
+        if (!spotManager.getUser().getId().equals(userId) && !user.getRole().equals(UserRole.ADMIN)) {
+            throw new UserIsNotAdminOrOwnerException("Você não tem autorização para isso");
+        }
 
         touristSpotRepository.delete(touristSpot);
+    }
+
+    private boolean timeIsValid(LocalTime opensAt, LocalTime closesAt) {
+        if (opensAt != null && closesAt != null) {
+            return !closesAt.isBefore(opensAt);
+        }
+        return true;
     }
 }
