+++ backend/src/main/java/io/turismo/backend/service/WarnService.java
@@ -2,6 +2,8 @@ package io.turismo.backend.service;
 
 import io.turismo.backend.dto.warn.WarnCreateDTO;
 import io.turismo.backend.dto.warn.WarnDTO;
+import io.turismo.backend.exception.ObjectNotFoundException;
+import io.turismo.backend.exception.UserIsNotOwnerException;
 import io.turismo.backend.mapper.WarnMapper;
 import io.turismo.backend.model.TouristSpot;
 import io.turismo.backend.model.Warn;
@@ -11,6 +13,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.UUID;
 
 @Service
@@ -25,12 +28,17 @@ public class WarnService {
         this.touristSpotRepository = touristSpotRepository;
     }
 
-    public WarnDTO createWarn(WarnCreateDTO dto, UUID touristSpotId){
+    public WarnDTO createWarn(UUID userSpotManagerId, WarnCreateDTO dto, UUID touristSpotId){
         TouristSpot touristSpot = touristSpotRepository.findById(touristSpotId)
-                .orElseThrow(() -> new RuntimeException("Ponto turístico não encontrado"));
+                .orElseThrow(() -> new ObjectNotFoundException("Ponto turístico não encontrado"));
+
+        if(!touristSpot.getSpotManager().getUser().getId().equals(userSpotManagerId)){
+            throw new UserIsNotOwnerException("Você não tem autorização para isso");
+        }
 
         Warn warn = warnMapper.toEntity(dto);
         warn.setTouristSpot(touristSpot);
+        warn.setEventDate(LocalDate.now());
 
         return warnMapper.toDTO(warnRepository.save(warn));
     }
@@ -38,7 +46,7 @@ public class WarnService {
     public WarnDTO getWarn(UUID warnId){
         return warnMapper.toDTO(
                 warnRepository.findById(warnId)
-                .orElseThrow(() -> new RuntimeException("Aviso não encontrado"))
+                .orElseThrow(() -> new ObjectNotFoundException("Aviso não encontrado"))
         );
     }
 
@@ -47,9 +55,14 @@ public class WarnService {
                 .map(warnMapper::toDTO);
     }
 
-    public void deleteWarn(UUID warnId) {
+    public void deleteWarn(UUID userSpotManagerId, UUID warnId) {
         Warn warn = warnRepository.findById(warnId)
-                        .orElseThrow(() -> new RuntimeException("Aviso não encontrado"));
+                        .orElseThrow(() -> new ObjectNotFoundException("Aviso não encontrado"));
+        TouristSpot touristSpot = warn.getTouristSpot();
+
+        if(!touristSpot.getSpotManager().getUser().getId().equals(userSpotManagerId)){
+            throw new UserIsNotOwnerException("Você não tem autorização para isso");
+        }
 
         warnRepository.delete(warn);
     }
